+++ src/pages/auths/authReset/AuthReset.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+import * as S from "./style";
+function AuthReset() {
+  return <div>AuthReset</div>;
+}
+
+export default AuthReset;
