+++ src/pages/auths/authSignup/AuthSignup.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+import * as S from "./style";
+function Signup() {
+  return <div>Signup</div>;
+}
+
+export default Signup;
