+++ src/pages/auths/authSignup/style.jsx
@@ -1,0 +1,1 @@
+import { styled } from "styled-components";
